Cloned LoRa code from StratoCore_RACHUTS
The RACHUTS source code is a work in progress, but this gets me started with RATS
LoRa. The RACHUTS code is for a LoRa connection to the PU, so the
message handling is probably not correct.

Because LoRa is not based on serial interface, we will not use SerialComm.
It is interrupt driven, a buffer is polled for new messages.

In RACHUTS, the LoRaRX() function is called from InstrumentLoop().
Would it fit better in the system to follow the "router" pattern, similar
to the method used with MCBRouter?